refactor: unify gateway and executor legacy parser code
