+++ jina/jaml/parsers/base.py
@@ -1,4 +1,7 @@
-from typing import TYPE_CHECKING, Any, Dict, Optional, Union
+import inspect
+from abc import ABC
+from functools import reduce
+from typing import TYPE_CHECKING, Any, Dict, Optional, Set, Type, Union
 
 if TYPE_CHECKING:
     from jina.orchestrate.flow.base import Flow
@@ -37,3 +40,49 @@ class VersionedYAMLParser:
         :param data: versioned flow object
         """
         raise NotImplementedError
+
+
+class BaseLegacyParser(VersionedYAMLParser, ABC):
+    """
+    BaseLegacyParser for classes that need parameter injection and that will be managed inside a runtime
+    for instance, :class:`BaseExecutor` and :class:`BaseGateway`
+    """
+
+    @staticmethod
+    def _get_all_arguments(class_):
+        """
+
+        :param class_: target class from which we want to retrieve arguments
+        :return: all the arguments of all the classes from which `class_` inherits
+        """
+
+        def get_class_arguments(class_):
+            """
+            :param class_: the class to check
+            :return: a list containing the arguments from `class_`
+            """
+            signature = inspect.signature(class_.__init__)
+            class_arguments = [p.name for p in signature.parameters.values()]
+            return class_arguments
+
+        def accumulate_classes(cls) -> Set[Type]:
+            """
+            :param cls: the class to check
+            :return: all classes from which cls inherits from
+            """
+
+            def _accumulate_classes(c, cs):
+                cs.append(c)
+                if cls == object:
+                    return cs
+                for base in c.__bases__:
+                    _accumulate_classes(base, cs)
+                return cs
+
+            classes = []
+            _accumulate_classes(cls, classes)
+            return set(classes)
+
+        all_classes = accumulate_classes(class_)
+        args = list(map(lambda x: get_class_arguments(x), all_classes))
+        return set(reduce(lambda x, y: x + y, args))
