+++ jina/jaml/parsers/gateway/legacy.py
@@ -1,55 +1,11 @@
-import inspect
-from functools import reduce
-from typing import Any, Dict, Optional, Set, Type
+from typing import Any, Dict, Optional, Type
 
-from jina.jaml.parsers.base import VersionedYAMLParser
+from jina.jaml.parsers.base import BaseLegacyParser
 from jina.serve.gateway import BaseGateway
 
 
-# TODO: unify code with jina/jaml/parsers/executor/legacy.py
-class LegacyParser(VersionedYAMLParser):
+class GatewayLegacyParser(BaseLegacyParser):
     """Legacy parser for gateway."""
-
-    version = 'legacy'  # the version number this parser designed for
-
-    @staticmethod
-    def _get_all_arguments(class_):
-        """
-
-        :param class_: target class from which we want to retrieve arguments
-        :return: all the arguments of all the classes from which `class_` inherits
-        """
-
-        def get_class_arguments(class_):
-            """
-            :param class_: the class to check
-            :return: a list containing the arguments from `class_`
-            """
-            signature = inspect.signature(class_.__init__)
-            class_arguments = [p.name for p in signature.parameters.values()]
-            return class_arguments
-
-        def accumulate_classes(cls) -> Set[Type]:
-            """
-            :param cls: the class to check
-            :return: all classes from which cls inherits from
-            """
-
-            def _accumulate_classes(c, cs):
-                cs.append(c)
-                if cls == object:
-                    return cs
-                for base in c.__bases__:
-                    _accumulate_classes(base, cs)
-                return cs
-
-            classes = []
-            _accumulate_classes(cls, classes)
-            return set(classes)
-
-        all_classes = accumulate_classes(class_)
-        args = list(map(lambda x: get_class_arguments(x), all_classes))
-        return set(reduce(lambda x, y: x + y, args))
 
     def parse(
         self,
